Compiler warning fixes (#870)
* added cmake presets for enabling additional compiler warnings (as in VC)
* basic fixes
-removed superfluous semicolon
- added static_cast where it will definitely cause no change in behaviour
- small adaptions in member and parameter types

+++ examples/example.cpp
@@ -155,7 +155,7 @@ int main()
           auto x = crow::json::load(req.body);
           if (!x)
               return crow::response(400);
-          int sum = x["a"].i() + x["b"].i();
+          int64_t sum = x["a"].i() + x["b"].i();
           std::ostringstream os;
           os << sum;
           return crow::response{os.str()};
